requirement.txt
streamlit
pandas
scikit-learn
googletrans==4.0.0-rc1
nltk
sentence-transformers
faiss-cpu
numpy
